% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{coolor_colour}
\alias{coolor_colour}
\title{Convert a coolors.co URL into a vector of hex codes}
\usage{
coolor_colour(cooler_url)
}
\arguments{
\item{cooler_url}{}
}
\value{
a vector of hex codes
}
\description{
Convert a coolors.co URL into a vector of hex codes
}
