% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{get_all_colours}
\alias{get_all_colours}
\title{Get all available colour palettes}
\usage{
get_all_colours()
}
\value{
a named list containing all available colour palettes
}
\description{
Get all available colour palettes
}
