% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{get_colour_names}
\alias{get_colour_names}
\title{Get just the names of all available colour palettes}
\usage{
get_colour_names()
}
\value{
A vector of palette names that are available
}
\description{
Get just the names of all available colour palettes
}
